Added libs and moved components to directory
