+++ new_architecture/package.json
@@ -18,7 +18,12 @@
         "eslint:fix": "npm run eslint -- --fix"
     },
     "dependencies": {
-        "node-sass": "^7.0.1",
+        "@hookform/resolvers": "^2.8.5",
+        "@hqoss/monads": "^0.5.0",
+        "class-transformer": "^0.5.1",
+        "class-validator": "^0.13.2",
+        "ky": "^0.28.7",
+        "njct": "^8.1.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-scripts": "5.0.0",
