+++ new_architecture/src/components/App/App.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared';
+@import '../../shared/shared';
 
 body {
     margin: 0;
