+++ new_architecture/src/components/Article/Article.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared';
+@import '../../shared/shared';
 
 $light-mode-vote-color: rgba($color: $light-mode-color, $alpha: .6);
 $dark-mode-vote-color: rgba($color: $dark-mode-color, $alpha: .6);
