+++ new_architecture/src/components/ArticleList/ArticleList.component.tsx
@@ -4,8 +4,8 @@ import React, { useEffect, useState } from 'react';
 import { useRecoilValueLoadable } from 'recoil';
 
 import Article from '../Article/Article.component';
-import ArticleService from '../shared/ArticleService';
-import { articleByIdQuery } from '../shared/GlobalState';
+import ArticleService from '../../shared/ArticleService';
+import { articleByIdQuery } from '../../shared/GlobalState';
 
 const ArticleList = () => {
   const article = useRecoilValueLoadable(articleByIdQuery(3));
