+++ new_architecture/src/components/ArticleList/ArticleList.style.scss
