+++ new_architecture/src/components/Breadcrumbs/Breadcrumbs.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared';
+@import '../../shared/shared';
 
 .Breadcrumbs {
     align-self: flex-start;
