+++ new_architecture/src/components/Header/Header.style.scss
@@ -1,4 +1,4 @@
-@import '../shared/shared';
+@import '../../shared/shared';
 
 .Header {
     width: 100%;
