+++ new_architecture/src/components/MainContent/MainContent.component.tsx
