+++ new_architecture/src/components/MainContent/MainContent.style.scss
