+++ new_architecture/tsconfig.json
@@ -1,58 +1,40 @@
 {
     "ts-node": {
-        "transpileOnly": true,
-        "compilerOptions": {
-            "target": "es2018",
-            "module": "commonjs"
-        }
+      "transpileOnly": true,
+      "compilerOptions": {
+        "target": "es2018",
+        "module": "commonjs"
+      }
     },
     "compilerOptions": {
-        "target": "esnext",
-        "module": "esnext",
-        "lib": [
-            "dom",
-            "esnext"
-        ],
-        "allowJs": false,
-        "noUncheckedIndexedAccess": true,
-        "skipLibCheck": true,
-        "esModuleInterop": true,
-        "allowSyntheticDefaultImports": true,
-        "strict": true,
-        "forceConsistentCasingInFileNames": true,
-        "moduleResolution": "node",
-        "resolveJsonModule": true,
-        "isolatedModules": true,
-        "experimentalDecorators": true,
-        "useUnknownInCatchVariables": true,
-        "noEmit": true,
-        "jsx": "react",
-        "baseUrl": ".",
-        "paths": {
-            "@libs/application": [
-                "src/application/index.ts"
-            ],
-            "@libs/application/*": [
-                "src/application/*"
-            ],
-            "@libs/components": [
-                "src/components/index.ts"
-            ],
-            "@libs/components/*": [
-                "src/components/*"
-            ],
-            "@libs/api": [
-                "src/api/index.ts"
-            ],
-            "@libs/api/*": [
-                "src/api/*"
-            ],
-            "@app_modules/*": [
-                "src/app_modules/*"
-            ]
-        }
+      "target": "esnext",
+      "module": "esnext",
+      "lib": ["dom", "esnext"],
+      "allowJs": false,
+      "noUncheckedIndexedAccess": true,
+      "skipLibCheck": true,
+      "esModuleInterop": true,
+      "allowSyntheticDefaultImports": true,
+      "strict": true,
+      "forceConsistentCasingInFileNames": true,
+      "moduleResolution": "node",
+      "resolveJsonModule": true,
+      "isolatedModules": true,
+      "experimentalDecorators": true,
+      "useUnknownInCatchVariables": true,
+      "noEmit": true,
+      "jsx": "react",
+      "baseUrl": ".",
+      "paths": {
+        "@libs/application": ["src/application/index.ts"],
+        "@libs/application/*": ["src/application/*"],
+        "@libs/components": ["src/components/index.ts"],
+        "@libs/components/*": ["src/components/*"],
+        "@libs/api": ["src/api/index.ts"],
+        "@libs/api/*": ["src/api/*"],
+        "@app_modules/*": ["src/app_modules/*"]
+      }
     },
-    "include": [
-        "src"
-    ]
-}
+    "include": ["src"]
+  }
+
